[Locus] add PseudoLocus and change board representation
Add a new type PseudoLocus that are locaii in the 0x88 board
representation. These can't index into a Board anymore, that required a
Locus. We do this so that we can use the smallest board representation,
a 64 element array, rather than a 0x88 one. Since all attacks and moves
are precomputed, the overhead of this tralsation between locaii formats
is minimal.

Elo difference: 3.47 +/- 23.31